Spring/jpa/src/main/java/me/minho/jpastudy/service/OrderService.java
package me.minho.jpastudy.service;

import lombok.AllArgsConstructor;
import me.minho.jpastudy.domain.*;
import me.minho.jpastudy.repository.ItemRepository;
import me.minho.jpastudy.repository.MemberRepository;
import me.minho.jpastudy.repository.OrderRepository;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

/**
 * IDE : IntelliJ IDEA
 * Created by minho on 2019-07-09.
 */

@Service
@Transactional
@AllArgsConstructor
public class OrderService {

    private OrderRepository orderRepository;
    private MemberRepository memberRepository;
    private ItemRepository itemRepository;

    public Long order(Long memberId, Long itemId, int count) {
        Member member = memberRepository.findOne(memberId);
        Item item = itemRepository.findOne(itemId);
        Delivery delivery = new Delivery(member.getAddress());

        OrderItem orderItem = OrderItem.createOrderItem(item, item.getPrice(), count);

        Order order = Order.createOrder(member, delivery, orderItem);

        orderRepository.save(order);

        return order.getId();
    }

    public void cancel(Long orderId) {
        Order order = orderRepository.findOne(orderId);
        order.cancel();
    }

    public List<Order> findOrders(OrderSearch orderSearch) {
        return orderRepository.findAll(orderSearch);
    }
}
